Get rid of unqualified size_t's in Chapter 5 exercises

+++ Exercises/Modules/Chapter 05/Soln5_06.cpp
@@ -14,8 +14,8 @@ int main()
   for (unsigned i {1}; i <= bound; ++i)
     values.push_back(i);
 
-  size_t count {};              // Number of output values
-  size_t perline {10};          // Number output perline                 
+  unsigned count {};              // Number of output values
+  unsigned perline {10};          // Number output perline                 
   for (auto value : values)
   {
     if (value % 7 == 0 || value % 13 == 0) continue;
